fix: Add explicit type casting to resolve mypy errors
Fixed 8 mypy type checking errors for CI/CD pipeline:

1. sensor.py:624 - Cast round() to int for ping latency
2. sensor.py:646 - Cast round() to int for HTTP latency
3. fallback.py:200 - Cast title to str
4. fallback.py:210 - Cast content to str
5. button.py:90 - Add return type annotation and cast to bool
6. test_fallback.py:250 - Add isinstance() check before .lower()
7. test_modem_scraper.py:426 - Add list[type] annotation
8. test_modem_scraper.py:451 - Add list[type] annotation

All changes maintain existing functionality while satisfying
strict type checking requirements.

+++ custom_components/cable_modem_monitor/sensor.py
@@ -621,7 +621,7 @@ class ModemPingLatencySensor(ModemSensorBase):
         ping_latency = self.coordinator.data.get("ping_latency_ms")
         if ping_latency is None:
             return None
-        return round(ping_latency)
+        return int(round(ping_latency))
 
 
 class ModemHttpLatencySensor(ModemSensorBase):
@@ -643,4 +643,4 @@ class ModemHttpLatencySensor(ModemSensorBase):
         http_latency = self.coordinator.data.get("http_latency_ms")
         if http_latency is None:
             return None
-        return round(http_latency)
+        return int(round(http_latency))
